Add products with image

+++ backend/src/main/java/com/arpan/backend/entity/Products.java
@@ -20,7 +20,7 @@ import java.util.Date;
 public class Products {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int prodId;
+    private Long prodId;
     private int stockQuantity;
     private BigDecimal price;
     private String name;
