+++ backend/src/main/java/com/arpan/backend/repository/ProductRepo.java
@@ -3,6 +3,6 @@ package com.arpan.backend.repository;
 import com.arpan.backend.entity.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepo extends JpaRepository<Products, Integer> {
+public interface ProductRepo extends JpaRepository<Products, Long> {
 
 }
